backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
dependencies = [ "flask>=3.1, <4", "flask-cors>=5, <6", "python-dotenv>=1.0, <2", "requests>=2.32, <3", "pypdf>=5.4, <6", "python-docx>=1.1, <2", "pydantic>=2.10, <3", "langchain>=0.3.25, <0.4", "langchain-community>=0.3.24, <0.4", "dashscope>=1.23, <2", "alibabacloud-bailian20231229>=2.7.0", "alibabacloud-tea-openapi>=0.3.15", "alibabacloud-tea-util>=0.3.13", "pytest>=8.3, <9" ]
requires-python = "~=3.13.0"

[tool.vercel.scripts]
build = "python build.py"
